+++ x/btcstaking/keeper/genesis.go
@@ -62,6 +62,26 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 		}
 	}
 
+	for _, fpAddrStr := range gs.FpBbnAddr {
+		fpAddr, err := sdk.AccAddressFromBech32(fpAddrStr)
+		if err != nil {
+			return fmt.Errorf("error decoding fp addr %s: %w", fpAddrStr, err)
+		}
+		if err := k.SetFpBbnAddr(ctx, fpAddr); err != nil {
+			return err
+		}
+	}
+
+	for _, deletedFpBtcPkHex := range gs.DeletedFpsBtcPkHex {
+		deletedFpBtcPk, err := bbn.NewBIP340PubKeyFromHex(deletedFpBtcPkHex)
+		if err != nil {
+			return fmt.Errorf("error decoding blocked fp btc pk hex %s: %w", deletedFpBtcPkHex, err)
+		}
+		if err := k.SoftDeleteFinalityProvider(ctx, deletedFpBtcPk); err != nil {
+			return err
+		}
+	}
+
 	if gs.LargestBtcReorg != nil {
 		if err := k.SetLargestBtcReorg(ctx, *gs.LargestBtcReorg); err != nil {
 			return err
@@ -97,15 +117,27 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
+	fpBbnAddr, err := k.fpBtcPkByAddrEntries(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	deletedFps, err := k.deletedFps(ctx)
+	if err != nil {
+		return nil, err
+	}
+
 	return &types.GenesisState{
-		Params:            k.GetAllParams(ctx),
-		FinalityProviders: fps,
-		BtcDelegations:    dels,
-		BlockHeightChains: k.blockHeightChains(ctx),
-		BtcDelegators:     btcDels,
-		Events:            evts,
-		LargestBtcReorg:   k.GetLargestBtcReorg(ctx),
-		ConsumerEvents:    k.consumerEvents(ctx),
+		Params:                      k.GetAllParams(ctx),
+		FinalityProviders:           fps,
+		BtcDelegations:              dels,
+		BlockHeightChains:           k.blockHeightChains(ctx),
+		BtcDelegators:               btcDels,
+		Events:                      evts,
+		LargestBtcReorg:             k.GetLargestBtcReorg(ctx),
+		ConsumerEvents:              k.consumerEvents(ctx),
+		FpBbnAddr:                   fpBbnAddr,
+		DeletedFpsBtcPkHex:          deletedFps,
 	}, nil
 }
 
@@ -114,12 +146,12 @@ func (k Keeper) finalityProviders(ctx context.Context) ([]*types.FinalityProvide
 	iter := k.finalityProviderStore(ctx).Iterator(nil, nil)
 	defer iter.Close()
 
-	for ; iter.Valid(); iter.Next() {
-		var fp types.FinalityProvider
-		if err := fp.Unmarshal(iter.Value()); err != nil {
-			return nil, err
-		}
+	err := k.IterateFinalityProvider(ctx, func(fp types.FinalityProvider) error {
 		fps = append(fps, &fp)
+		return nil
+	})
+	if err != nil {
+		return nil, err
 	}
 
 	return fps, nil
@@ -274,6 +306,47 @@ func (k Keeper) setConsumerEvents(ctx context.Context, events []*types.ConsumerE
 		store.Set(storeKey, eventsBytes)
 	}
 	return nil
+}
+
+func (k Keeper) deletedFps(ctx context.Context) ([]string, error) {
+	entries := make([]string, 0)
+	err := k.finalityProvidersDeleted.Walk(ctx, nil, func(key []byte) (stop bool, err error) {
+		fpBtcPk, err := bbn.NewBIP340PubKey(key)
+		if err != nil {
+			return true, err
+		}
+
+		entries = append(entries, fpBtcPk.MarshalHex())
+		return false, nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return entries, nil
+}
+
+func (k Keeper) fpBtcPkByAddrEntries(ctx context.Context) ([]string, error) {
+	entries := make([]string, 0)
+
+	iterator, err := k.fpBbnAddr.Iterate(ctx, nil)
+	if err != nil {
+		return nil, err
+	}
+	defer iterator.Close()
+
+	for ; iterator.Valid(); iterator.Next() {
+		key, err := iterator.Key()
+		if err != nil {
+			return nil, err
+		}
+		if len(key) == 0 {
+			continue
+		}
+		entries = append(entries, sdk.AccAddress(key).String())
+	}
+
+	return entries, nil
 }
 
 // parseUintsFromStoreKey expects to receive a key with