+++ x/btcstaking/keeper/grpc_query.go
@@ -41,7 +41,9 @@ func (k Keeper) FinalityProviders(c context.Context, req *types.QueryFinalityPro
 		if err != nil {
 			return err
 		}
-		resp := types.NewFinalityProviderResponse(fp, currBlockHeight)
+
+		isDeleted := k.IsFinalityProviderDeleted(ctx, fp.BtcPk)
+		resp := types.NewFinalityProviderResponse(fp, currBlockHeight, isDeleted)
 		fpResp = append(fpResp, resp)
 		return nil
 	})
@@ -82,7 +84,8 @@ func (k Keeper) FinalityProvider(c context.Context, req *types.QueryFinalityProv
 		return nil, err
 	}
 
-	fpResp := types.NewFinalityProviderResponse(fp, currBlockHeight)
+	isDeleted := k.IsFinalityProviderDeleted(ctx, fp.BtcPk)
+	fpResp := types.NewFinalityProviderResponse(fp, currBlockHeight, isDeleted)
 	return &types.QueryFinalityProviderResponse{FinalityProvider: fpResp}, nil
 }