+++ x/btcstaking/types/genesis.go
@@ -7,9 +7,11 @@ import (
 	"fmt"
 	"sort"
 
+	bbn "github.com/babylonlabs-io/babylon/v4/types"
 	types "github.com/babylonlabs-io/babylon/v4/types"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/cosmos/cosmos-sdk/codec"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // DefaultGenesis returns the default genesis state
@@ -82,9 +84,11 @@ func (gs GenesisState) Validate() error {
 		return err
 	}
 
-	return types.ValidateEntries(gs.ConsumerEvents, func(e *ConsumerEvent) string {
-		return e.ConsumerId
-	})
+	if err := gs.validateFpBbnAddr(gs.FinalityProviders); err != nil {
+		return err
+	}
+
+	return gs.validateDeletedFps(gs.FinalityProviders)
 }
 
 // GenesisStateFromAppState returns x/btcstaking GenesisState given raw application
@@ -110,6 +114,48 @@ func (h AllowedStakingTxHashStr) Validate() error {
 	// NewHash validates hash size
 	if _, err := chainhash.NewHash(bz); err != nil {
 		return err
+	}
+	return nil
+}
+
+// validateDeletedFps validates there is only valid blocked fp btc pk
+func (gs GenesisState) validateDeletedFps(fps []*FinalityProvider) error {
+	mapFpBtcPk := make(map[string]struct{})
+	for _, fp := range fps {
+		mapFpBtcPk[fp.BtcPk.MarshalHex()] = struct{}{}
+	}
+
+	for _, deletedFpBtcPkHex := range gs.DeletedFpsBtcPkHex {
+		_, exist := mapFpBtcPk[deletedFpBtcPkHex]
+		if !exist {
+			return fmt.Errorf("fp btc pk %s is not in the fp list", deletedFpBtcPkHex)
+		}
+
+		_, err := bbn.NewBIP340PubKeyFromHex(deletedFpBtcPkHex)
+		if err != nil {
+			return fmt.Errorf("error decoding deleted fp btc pk hex %s: %w", deletedFpBtcPkHex, err)
+		}
+	}
+	return nil
+}
+
+// validateFpBbnAddr validates there is no duplicate fp bbn addr
+func (gs GenesisState) validateFpBbnAddr(fps []*FinalityProvider) error {
+	mapFpAddr := make(map[string]struct{})
+	for _, fp := range fps {
+		mapFpAddr[fp.Addr] = struct{}{}
+	}
+
+	for _, fpAddr := range gs.FpBbnAddr {
+		_, exist := mapFpAddr[fpAddr]
+		if !exist {
+			return fmt.Errorf("fp addr %s is not in the fp list", fpAddr)
+		}
+
+		_, err := sdk.AccAddressFromBech32(fpAddr)
+		if err != nil {
+			return fmt.Errorf("error decoding fp addr %s: %w", fpAddr, err)
+		}
 	}
 	return nil
 }
@@ -253,6 +299,14 @@ func SortData(gs *GenesisState) {
 
 	sort.Slice(gs.ConsumerEvents, func(i, j int) bool {
 		return gs.ConsumerEvents[i].ConsumerId < gs.ConsumerEvents[j].ConsumerId
+	})
+
+	sort.Slice(gs.FpBbnAddr, func(i, j int) bool {
+		return gs.FpBbnAddr[i] < gs.FpBbnAddr[j]
+	})
+
+	sort.Slice(gs.DeletedFpsBtcPkHex, func(i, j int) bool {
+		return gs.DeletedFpsBtcPkHex[i] < gs.DeletedFpsBtcPkHex[j]
 	})
 }